requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Django==5.0.4
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.1
pillow==10.4.0
python-pptx==1.0.2
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.4
XlsxWriter==3.2.0
